prometheus = "0.13.0" upgrade

+++ crates/util/stats/src/lib.rs
@@ -50,7 +50,8 @@ impl PrometheusRegistry {
         let name = format!("{}{}", self.prefix, name);
         let c = prometheus::IntCounter::new(name.as_str(), help)
             .expect("name and help must be non-empty");
-        c.inc_by(value);
+        c.inc_by(value as u64);
+        
         self.registry
             .register(Box::new(c))
             .expect("prometheus identifiers must be unique");
